extracted methods from user.js to note.js
