+++ src/routes/cohort.js
@@ -10,5 +10,6 @@ const router = Router()
 router.post('/', validateAuthentication, validateTeacherRole, create)
 router.get('/', validateAuthentication, validateTeacherRole, cohorts)
 router.get('/:id', validateAuthentication, validateTeacherRole, getCohortById)
+// maybe router.get("/:id/cohortExercises") to get all the cohort exercises from a specific cohort
 
 export default router
